requirements.txt
argparse
dnspython
requests
lxml
yapf